test: build the integration suite as one binary
Cargo links a separate binary per test target, and `duckdb-bundled` is a default
feature, so each of the ~50 `tests/*.rs` files statically linked its own copy of
DuckDB — 260-430 MB apiece, several gigabytes of linker output per build. That is
what exhausted the runner's disk while linking (`lld` failing with SIGBUS, which
`CARGO_PROFILE_TEST_DEBUG: "0"` in CI works around) and what inflated the cached
target directory past GitHub's 10 GB per-repository budget: after #227 the macOS
entry alone reached 5.3 GB, and saving it evicted every other cache in the
repository.

Every former `tests/*.rs` becomes a module of `tests/it/main.rs`, declared behind
`autotests = false` and an explicit `[[test]] name = "it"`. One binary, 371 MB.

Mechanically this is a rename plus two edits per affected file:

- The 50 `#![cfg(feature = ...)]` inner attributes are untouched. An inner `cfg`
  on a module file compiles the module away when the feature is off, exactly as
  the standalone binary previously compiled to zero tests.
- `mod common;` becomes `use crate::common;` (14 files) and `mod hybrid_asyncdb;`
  becomes `use crate::hybrid_asyncdb;` (1 file), so every existing `common::foo`
  call site is unchanged. Three files declared `mod common;` without using it;
  as an import that is a warning, and CI lints with `-D warnings`, so they were
  dropped.
- `tests/sqllogictests/` stays put — the runner resolves it relative to the
  package root, not to its own source location.

Verified against 8dc58ac by diffing `cargo nextest list`: 611 tests before, 609
after, nothing gained, and the two absent entries are
`sqllogictest_runner::hybrid_asyncdb::tests::{test_table_rewrite,
test_write_detection}` — duplicate executions, because `hybrid_asyncdb.rs` was
both its own binary and a module of the runner. Both still run once under
`hybrid_asyncdb::tests`. So no coverage is lost. Suite green (609 passed) and
`cargo clippy --workspace --all-targets --all-features -- -D warnings` clean,
which is what compiles the postgres/mysql/multicatalog modules a default-feature
build skips.

A test's name now carries its module as a prefix
(`write_tests::test_append_semantics`). Substring filters are unaffected, so the
documented `cargo test delete_filter` still works — better than before, in fact,
since it previously matched test names only and no test in that file contains
that string. Targeting one former file is now `cargo test --test it <module>::`
rather than `--test <file>`; `tests/README.md` documented the old form five
times and is updated, along with stale `tests/<file>.rs` links in `README.md`,
`CLAUDE.md`, and three doc comments.
