+++ tests/it/append_with_deletes_tests.rs
