+++ tests/it/cdc_cumulative_delete_tests.rs
