+++ tests/it/inlined_data_sqlite_tests.rs
