+++ tests/it/multicatalog_hardening_tests.rs
