+++ tests/it/numeric_metadata_validation_tests.rs
