+++ tests/it/positional_delete_oracle_tests.rs
