+++ tests/it/renamed_columns_tests.rs
