+++ tests/it/table_deletions_repartition_tests.rs
@@ -10,7 +10,6 @@
 //! missed, and positions matched in the wrong partition emitted the wrong
 //! row's content. All tests here FAIL on the pre-fix code.
 
-mod common;
 
 use std::collections::BTreeSet;
 use std::path::Path;
